refactor: align sync channel message query timeout handling
